backend/requirements.txt
flask
flask-cors
PyPDF2
groq
requests
python-dotenv